requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
pyyaml>=6.0

# Deep Learning
torch>=1.13.0
torchvision>=0.14.0
torchaudio>=0.13.0

# Utilities
tqdm>=4.62.0
pillow>=9.0.0
ipykernel>=6.0.0
jupyter>=1.0.0
notebook>=6.4.0

# Optional - for experiment tracking
# tensorboard>=2.8.0
# wandb>=0.12.0

# Optional - for efficient loading
# albumentations>=1.0.0
